Grouping extensions for deadline, deadline color and deadline type
