+++ app/src/main/java/com/kevalpatel2106/yip/dashboard/DashboardActivity.kt
@@ -11,11 +11,11 @@ import androidx.navigation.navArgs
 import com.google.android.gms.ads.AdRequest
 import com.google.android.gms.ads.InterstitialAd
 import com.kevalpatel2106.yip.R
-import com.kevalpatel2106.yip.core.getLaunchIntent
+import com.kevalpatel2106.yip.core.ext.getLaunchIntent
+import com.kevalpatel2106.yip.core.ext.showSnack
+import com.kevalpatel2106.yip.core.ext.slideDown
+import com.kevalpatel2106.yip.core.ext.slideUp
 import com.kevalpatel2106.yip.core.livedata.nullSafeObserve
-import com.kevalpatel2106.yip.core.showSnack
-import com.kevalpatel2106.yip.core.slideDown
-import com.kevalpatel2106.yip.core.slideUp
 import com.kevalpatel2106.yip.dashboard.adapter.DeadlineAdapter
 import com.kevalpatel2106.yip.dashboard.adapter.DeadlineAdapterEventListener
 import com.kevalpatel2106.yip.dashboard.inAppReview.InAppReviewManager
