+++ app/src/main/java/com/kevalpatel2106/yip/edit/EditDeadlineViewModel.kt
@@ -10,14 +10,14 @@ import androidx.lifecycle.MutableLiveData
 import com.kevalpatel2106.yip.R
 import com.kevalpatel2106.yip.core.BaseViewModel
 import com.kevalpatel2106.yip.core.RxSchedulers
-import com.kevalpatel2106.yip.core.addTo
+import com.kevalpatel2106.yip.core.ext.addTo
+import com.kevalpatel2106.yip.core.ext.setToDayMax
+import com.kevalpatel2106.yip.core.ext.setToDayMin
 import com.kevalpatel2106.yip.core.livedata.SingleLiveEvent
 import com.kevalpatel2106.yip.core.livedata.modify
 import com.kevalpatel2106.yip.core.livedata.nullSafeValue
-import com.kevalpatel2106.yip.core.setToDayMax
-import com.kevalpatel2106.yip.core.setToDayMin
-import com.kevalpatel2106.yip.entity.getDeadlineColor
-import com.kevalpatel2106.yip.entity.isPreBuild
+import com.kevalpatel2106.yip.entity.ext.getDeadlineColor
+import com.kevalpatel2106.yip.entity.ext.isPreBuild
 import com.kevalpatel2106.yip.notifications.DeadlineNotificationReceiver
 import com.kevalpatel2106.yip.repo.alarmRepo.AlarmRepo
 import com.kevalpatel2106.yip.repo.billingRepo.BillingRepo
