+++ app/src/main/java/com/kevalpatel2106/yip/webviews/WebViewViewState.kt
@@ -1,6 +1,6 @@
 package com.kevalpatel2106.yip.webviews
 
-import com.kevalpatel2106.yip.core.emptyString
+import com.kevalpatel2106.yip.core.ext.emptyString
 
 internal data class WebViewViewState(val title: String, val linkUrl: String) {
     companion object {
