+++ app/src/test/java/com/kevalpatel2106/yip/edit/EditViewDeadlineSaveDeadlineTest.kt
@@ -1,12 +1,12 @@
 package com.kevalpatel2106.yip.edit
 
 import com.kevalpatel2106.testutils.getOrAwaitValue
-import com.kevalpatel2106.yip.core.setToDayMax
-import com.kevalpatel2106.yip.core.setToDayMin
+import com.kevalpatel2106.yip.core.ext.setToDayMax
+import com.kevalpatel2106.yip.core.ext.setToDayMin
 import com.kevalpatel2106.yip.entity.Deadline
 import com.kevalpatel2106.yip.entity.DeadlineColor
 import com.kevalpatel2106.yip.entity.DeadlineType
-import com.kevalpatel2106.yip.entity.isPreBuild
+import com.kevalpatel2106.yip.entity.ext.isPreBuild
 import com.kevalpatel2106.yip.notifications.DeadlineNotificationReceiver
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.anyOrNull
