+++ core/src/main/java/com/kevalpatel2106/yip/core/ext/ContextExtention.kt
@@ -1,15 +1,18 @@
-package com.kevalpatel2106.yip.core
+package com.kevalpatel2106.yip.core.ext
 
 import android.app.Activity
 import android.appwidget.AppWidgetManager
 import android.content.ActivityNotFoundException
 import android.content.Context
 import android.content.Intent
+import android.graphics.Color
+import android.graphics.drawable.GradientDrawable
 import android.net.Uri
 import android.os.Build
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.annotation.ColorInt
 import androidx.annotation.ColorRes
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
@@ -21,6 +24,11 @@ import com.kevalpatel2106.feature.core.R
 const val SNACK_BAR_DURATION = 2000L
 private const val NEW_TASK_INTENT_FLAG =
     Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK
+private const val HIV_TO_COLOR_DIMENSION = 3
+private const val GRADIENT_70 = 0.7F
+private const val GRADIENT_80 = 0.8F
+private const val GRADIENT_85 = 0.85F
+private const val GRADIENT_90 = 0.9F
 
 /**
  * Get the color from color res.
@@ -108,3 +116,22 @@ fun Context.sendMailToDev() {
 }
 
 fun Context.updateWidgets() = sendBroadcast(Intent(AppWidgetManager.ACTION_APPWIDGET_UPDATE))
+
+@ColorInt
+fun darkenColor(@ColorInt color: Int, factor: Float = 0.5f): Int {
+    return Color.HSVToColor(FloatArray(HIV_TO_COLOR_DIMENSION).apply {
+        Color.colorToHSV(color, this)
+        this[2] *= factor
+    })
+}
+
+fun Context.getBackgroundGradient(@ColorInt color: Int): GradientDrawable {
+    val dark70 = darkenColor(color, GRADIENT_70)
+    val dark80 = darkenColor(color, GRADIENT_80)
+    val dark85 = darkenColor(color, GRADIENT_85)
+    val dark90 = darkenColor(color, GRADIENT_90)
+
+    val colors = intArrayOf(dark70, dark80, dark85, dark90, dark90, color)
+    return GradientDrawable(GradientDrawable.Orientation.LEFT_RIGHT, colors)
+        .apply { cornerRadius = resources.getDimension(R.dimen.card_radius) }
+}
