+++ core/src/main/java/com/kevalpatel2106/yip/core/ext/DateExt.kt
@@ -1,4 +1,4 @@
-package com.kevalpatel2106.yip.core
+package com.kevalpatel2106.yip.core.ext
 
 import java.util.Calendar
 import java.util.Date
