+++ core/src/test/java/com/kevalpatel2106/yip/core/ext/RxExtKtTest.kt
@@ -1,4 +1,4 @@
-package com.kevalpatel2106.yip.core
+package com.kevalpatel2106.yip.core.ext
 
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
