+++ core/src/test/java/com/kevalpatel2106/yip/core/ext/ViewExtTests.kt
@@ -1,9 +1,6 @@
-package com.kevalpatel2106.yip.core
+package com.kevalpatel2106.yip.core.ext
 
 import android.content.Context
-import android.graphics.Color
-import android.graphics.drawable.GradientDrawable
-import android.os.Build
 import android.view.MenuItem
 import android.view.View
 import android.widget.ProgressBar
@@ -12,12 +9,10 @@ import androidx.core.view.isVisible
 import androidx.test.core.app.ApplicationProvider
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
-import org.junit.Assert.assertNotEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Before
-import org.junit.Ignore
 import org.junit.Test
 import org.junit.experimental.runners.Enclosed
 import org.junit.runner.RunWith
@@ -119,46 +114,6 @@ class ViewExtTests {
             val view = View(ApplicationProvider.getApplicationContext<Context>())
             view.showOrHide(false)
             assertTrue(!view.isVisible)
-        }
-
-        @Test
-        @Throws(Exception::class)
-        fun testDarkerColor() {
-            val factor = 0.5f
-
-            val originalColorArray = FloatArray(3)
-            Color.colorToHSV(Color.DKGRAY, originalColorArray)
-
-            val darkerColorArray = FloatArray(3)
-            Color.colorToHSV(darkenColor(Color.DKGRAY, factor), darkerColorArray)
-
-            assertEquals(darkerColorArray[2], originalColorArray[2] * factor)
-        }
-
-        @Test
-        @Ignore
-        @Throws(Exception::class)
-        fun testGetBackgroundGradient() {
-            val testColor = Color.BLACK
-            val gradientDrawable = ApplicationProvider.getApplicationContext<Context>()
-                .getBackgroundGradient(testColor)
-
-            // Check orientation
-            assertEquals(GradientDrawable.Orientation.LEFT_RIGHT, gradientDrawable.orientation)
-
-            // Check radius
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                assertNotEquals(0F, gradientDrawable.cornerRadius)
-
-                // Check colors
-                assertEquals(6, gradientDrawable.colors?.size)
-                assertEquals(darkenColor(testColor, 0.7f), gradientDrawable.colors!![0])
-                assertEquals(darkenColor(testColor, 0.8f), gradientDrawable.colors!![1])
-                assertEquals(darkenColor(testColor, 0.85f), gradientDrawable.colors!![2])
-                assertEquals(darkenColor(testColor, 0.9f), gradientDrawable.colors!![3])
-                assertEquals(darkenColor(testColor, 0.9f), gradientDrawable.colors!![4])
-                assertEquals(testColor, gradientDrawable.colors!![5])
-            }
         }
 
         @Test
