+++ entity/src/main/java/com/kevalpatel2106/yip/entity/ext/DeadlineTypeExt.kt
@@ -1,0 +1,6 @@
+package com.kevalpatel2106.yip.entity.ext
+
+import com.kevalpatel2106.yip.entity.DeadlineType
+
+fun DeadlineType.isPreBuild() = this != DeadlineType.CUSTOM
+fun DeadlineType.isRepeatable() = this != DeadlineType.CUSTOM
